Looks like basic methods done
