+++ backend/api/models/__init__.py
@@ -1,4 +1,5 @@
 from base_model import BaseModel
 from model_controller import ModelController
+from model_info import ModelInfo
 from . import exc
 from .types import id_type
